handle duplicate SA names
